requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6
django-filter>=24.2
psycopg2-binary>=2.9
python-dotenv>=1.0
whitenoise>=6.6
httpx>=0.27
pytest>=8.2
pytest-django>=4.8
pytest-mock>=3.12
factory-boy>=3.3
faker>=19.0
